podia dar overflow antes :o

+++ lib/util.h
@@ -4,8 +4,8 @@
 #define INIT_DATA(X)                                                           \
   Data X = {.numberOfComparisons = 0, .numberOfSwaps = 0, .time = 0}
 typedef struct data_ {
-  int numberOfComparisons;
-  int numberOfSwaps;
+  unsigned long long numberOfComparisons;
+  unsigned long long numberOfSwaps;
   clock_t start, end;
   double time;
 } Data;
